refactor(frontend): split llm-api into lib/llm module
Apply the same split as lib/chat: lib/llm-api.ts becomes
lib/llm/{types,client,index}.ts with a barrel re-export, so callers
import from @/lib/llm instead.
