+++ frontend/components/LLMConfigSelect.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { Select } from "@/components/ui/Select";
 import { Alert } from "@/components/ui/Alert";
 import { useAuth } from "@/lib/auth-context";
-import { fetchLLMConfigs, type LLMConfig } from "@/lib/llm-api";
+import { fetchLLMConfigs, type LLMConfig } from "@/lib/llm";
 
 interface LLMConfigSelectProps {
   value: number | undefined;
