+++ frontend/lib/tests/llm.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 
-import { fetchProviderCatalog } from "@/lib/llm-api";
+import { fetchProviderCatalog } from "@/lib/llm";
 
 describe("fetchProviderCatalog", () => {
   beforeEach(() => {
